% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmlTemplates-package.R
\docType{package}
\name{fmlTemplates}
\alias{fmlTemplates}
\alias{fmlTemplates-package}
\title{fmlTemplates Package}
\description{
fmlTemplates a package with utilities functions and templates for classes
}
\details{
This package provides templates for classes such as syllabus, course, chapter
group assesment, individual assesment, assement instructions etc.
}
\author{
FML \email{fml@example.org}
}
\keyword{internal}
