% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{insert_course}
\alias{insert_course}
\title{Insert course summary}
\usage{
insert_course(
  title,
  topic,
  details = NULL,
  author = NULL,
  id = NULL,
  format = "inline"
)
}
\arguments{
\item{title}{The title of the course.}

\item{topic}{The main topic corresponding to the current document.}

\item{details}{The details or the summary of the topic. Default NULL.}

\item{author}{The author of the course. Default NULL.}

\item{id}{The course id. Default NULL.}

\item{format}{Format of the output. Default "inline"}
}
\value{
A formatted outline of the course
}
\description{
Insert the formatted summary of the course
}
\details{
This function has been designed to quickly insert a course summary
in various places of the template.
This function is not supposed to be included in a file that will be distributed.
}
\examples{

insert_course("Introduction to Statistics", "Descriptive Statistics")
}
\author{
FML
}
