% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testStatCourse}
\alias{testStatCourse}
\title{Data example for a statistics course}
\format{
## `testStatCourse`
A named vector of length 2
\describe{
  \item{title}{title of the course}
  \item{topic}{topic of the course}
}
}
\source{
Content generated for this package
}
\usage{
testStatCourse
}
\description{
Generated data example
}
\keyword{datasets}
